Created angleSolver.c and angleSolver.h which take in the length of the three sides of a triangle and calculate the inside angles of it.
Three functions in angleSolver.c: analyzeAngles, findAngle and checkLengths.

analyzeAngles prints the angles of the triangle given by the user and uses findAngle to calculate them. It uses checkLengths to find if the lengths given by the user form a regular closed triangle, if they do not it returns a message stateing so.